Made it so doors cost scrap

+++ GameJam2020_GroupA_Game/Assets/Scripts/GameManager.cs
@@ -74,7 +74,7 @@ public class GameManager : MonoBehaviour
     {
         fuel += a;
     }
-    public void subtrScrap(int a)
+    public void subtrScrap(float a)
     {
         scrap -= a;
     }
